[ROUTE FINDING] A basic route finding option is now implemented. Simply click on one of the items in the list and a Snackbar tells you, what the route to the product is. This is not optimized by now, but gives a general idea of how routing would work.
[STORE CHOOSER]
A market detail activity is now displayed prior to entering a market. You can therefore choose the specific market you want to use for your shopping more precisely with a picture, name, address and opening hours.

The HomeActivity was finally removed.

A bug within the serialized file "ekz.bin" will bring up a dialog at the start of the app, asking whether to try again or ignore the error. You can then re-save a valid version.

The Up-Navigation was changed, so that you will always be redirected to the main activity. This prevents the app from crashing.

NOTE: Internet connection and granting rogue GPS permission is now strictly necessarry to run sort the shopping list.
There are known bugs when these permissions are missing